% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_boundary_points.R
\name{x3p_boundary_points}
\alias{x3p_boundary_points}
\title{Identify boundary of a 3d topographic scan in x3p format}
\usage{
x3p_boundary_points(x3p, sample)
}
\arguments{
\item{x3p}{topographic scan in x3p format}

\item{sample}{positive integer value specifying the sampling ratio:
every `sample` value in x and y direction will be included.
Higher values result in faster results but cruder assessments of the boundary.}
}
\value{
data frame of boundary points, variables are named `x` and `y`
}
\description{
Identify boundary of a 3d topographic scan in x3p format
}
\examples{
library(x3ptools)
x3p <- x3p_read(system.file("T1AW-LI-R4.x3p", package="wires"))
bounds <- x3p_boundary_points(x3p, 2)

library(ggplot2)
library(dplyr)
bounds \%>\% ggplot(aes(x = x, y = y)) + geom_point()
}
