% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_surface_polygon.R
\name{x3p_surface_polygon}
\alias{x3p_surface_polygon}
\title{Identify the inside surface area of a set of points}
\usage{
x3p_surface_polygon(x, y, concavity, center = NULL)
}
\arguments{
\item{x}{numeric vector of x positions}

\item{y}{numeric vector of y positions}

\item{concavity}{strictly positive parameter. Smaller values approach the
inside of the boundary
more closely. Large values concentrate on the inner center of the area}

\item{center}{numeric vector of length 2, consisting of (x,y) coordinate for
centering x and y values.
If center is NULL, the half ranges of x and y are used.}
}
\value{
dataframe of x and y positions describing the inside of the area
described by the input x and y, variable id describes the order of the points
}
\description{
Identify the inside surface area of a set of points
}
\examples{
library(x3ptools)
x3p <- x3p_read("~/Documents/CSAFE/Wirecutter/data/Aluminum Wires renamed/T1AW-LI-R4-B07.x3p")
bounds <- x3p_boundary_points(x3p, 10)
polygon <- x3p_surface_polygon(bounds$x, bounds$y, 1)

library(ggplot2)
library(dplyr)
bounds \%>\%
ggplot(aes(x = x, y = y)) + geom_point() +
  geom_polygon(data = polygon)

}
